Dreamhack/dreamschool/src/app/requirements.txt
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-WTF==0.15.1
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
pycryptodome==3.10.1
PyJWT==1.7.1
pytz==2021.1
SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
